% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dual_filter}
\alias{dual_filter}
\title{Select features using `top_p` or `threshold`.}
\usage{
dual_filter(x, top_p, threshold, cutoff, maximize)
}
\arguments{
\item{x}{a named numeric vector of scores per feature}

\item{top_p}{an integer specifying the number of top-performing features to
retain}

\item{threshold}{a numeric with percentile of top-performing features to
retain. For example, `threshold = 0.9` will only retain features that are
in the top 90th percentile. A smaller value of threshold will select
more features.}

\item{cutoff}{a numeric with the value that represents the cutoff in the
scores in `x` by which to retain/discard features.}

\item{maximize}{logical to indicate whether `top_p`, `threshold` and `cutoff`
are used to keep features where high scores = 'best' (maximize = TRUE) or
where low scores = 'best' (maximize = FALSE).}
}
\value{
character vector of feature names to exclude
}
\description{
Feature selection using either the `top_p` or `threshold` features OR
`cutoff` where cutoff refers to the absolute numeric value of the feature
importance scores.
}
\details{
`dual_filter` selects feature that are selected using either (`top_p`,
`threshold`) or `cutoff` or both. If top_p/threshold and cutoff are both used
then features are selected using OR. For example, if top_p selects features 1
& 2, and threshold selects features 1 & 3, then the selected features =
1,2,3.
}
\keyword{internal}
