% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{entropy}
\alias{entropy}
\title{Parameter functions for feature selection recipes}
\usage{
entropy(values = values_entropy)
}
\arguments{
\item{values}{A character string of possible values. See `values_entropy` for
possible values.}
}
\value{
A function with classes "qual_param" and "param"
}
\description{
Entropy-based feature selection methods can be applied using several methods
to calculate the entropy formula. `entropy` is for specifying the type of
entropy-based filter that is used.
}
\examples{
entropy('infogain')
}
