% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_select_aov.R, R/step_select_boruta.R,
%   R/step_select_carscore.R, R/step_select_fcbf.R, R/step_select_forests.R,
%   R/step_select_infgain.R, R/step_select_linear.R, R/step_select_mrmr.R,
%   R/step_select_relief.R, R/step_select_roc.R, R/step_select_tree.R,
%   R/step_select_vip.R, R/step_select_xtab.R
\name{required_pkgs.step_select_aov}
\alias{required_pkgs.step_select_aov}
\alias{required_pkgs.step_select_boruta}
\alias{required_pkgs.step_select_carscore}
\alias{required_pkgs.step_select_fcbf}
\alias{required_pkgs.step_select_forests}
\alias{required_pkgs.step_select_infgain}
\alias{required_pkgs.step_select_linear}
\alias{required_pkgs.step_select_mrmr}
\alias{required_pkgs.step_select_relief}
\alias{required_pkgs.step_select_roc}
\alias{required_pkgs.step_select_tree}
\alias{required_pkgs.step_select_vip}
\alias{required_pkgs.step_select_xtab}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_select_aov}(x, ...)

\method{required_pkgs}{step_select_boruta}(x, ...)

\method{required_pkgs}{step_select_carscore}(x, ...)

\method{required_pkgs}{step_select_fcbf}(x, ...)

\method{required_pkgs}{step_select_forests}(x, ...)

\method{required_pkgs}{step_select_infgain}(x, ...)

\method{required_pkgs}{step_select_linear}(x, ...)

\method{required_pkgs}{step_select_mrmr}(x, ...)

\method{required_pkgs}{step_select_relief}(x, ...)

\method{required_pkgs}{step_select_roc}(x, ...)

\method{required_pkgs}{step_select_tree}(x, ...)

\method{required_pkgs}{step_select_vip}(x, ...)

\method{required_pkgs}{step_select_xtab}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
