% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_select_fcbf.R
\name{step_select_fcbf}
\alias{step_select_fcbf}
\title{Fast Correlation Based Filter for Feature Selection}
\usage{
step_select_fcbf(
  recipe,
  ...,
  min_su = 0.025,
  outcome = NA,
  cutpoint = 0.5,
  features_retained = NA,
  role = NA,
  trained = FALSE,
  removals = NULL,
  skip = FALSE,
  id = rand_id("select_fcbf")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{Selector functions that specify which features should be
considered by the FCBF, e.g., all_numeric_predictors(), all_predictors().}

\item{min_su}{Minimum threshold for symmetrical uncertainty. Lower values
allow more features to be selected.}

\item{outcome}{Outcome variable used for filter selection. If there is only
one outcome variable in the recipe, it will automatically be detected. If
multiple outcome variables exist, the user should specify it.}

\item{cutpoint}{Quantile value (0-1) describing how to split numeric features
into binary nominal features. e.g. 0.5 = median split}

\item{features_retained}{Internal object that gives a record of which
features were retained after FCBF. Should not be specified by the user.}

\item{role}{Not used for this step since new variables are not created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{removals}{Feature columns that will be removed. Used internally and
should not be set by the user.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
bake()? While all operations are baked when prep() is run, some operations
may not be able to be conducted on new data (e.g. processing the outcome
variable(s)). Care should be taken when using skip = TRUE as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
Returns the recipe object, with step_select_fcbf added to the
  sequence of operations for this recipe.
}
\description{
step_select_fcbf takes a set of features and performs a fast correlation
based filter, resulting in a smaller subset of features being selected. The
number of features selected depends on the min_su threshold parameter (a
lower threshold selects more features).
}
\details{
step_select_fcbf takes a range of features (e.g. the full feature
  set) and selects a subset of features using the FCBF algorithm as described
  in Yu, L. and Liu, H. (2003).

FCBF selects features to simultaneously minimize correlation between features
and maximise correlations between the features and the target. FCBF only
works with categorical features, so continuous features must first be
discretized. By default this is based on a median split (i.e. splitting
continuous variables into 'high' versus 'low'), but the method may be
customized in the internal function 'discretize_var'.

Code to implement the FCBF algorithm is driven by Bioconductor package
'FCBF'. step_select_fcbf provides wrappers that allow it to be used within
the tidymodels framework
}
\references{
Yu, L. and Liu, H. (2003); Feature Selection for High-Dimensional
  Data A Fast Correlation Based Filter Solution, Proc. 20th Intl. Conf. Mach.
  Learn. (ICML-2003), Washington DC, 2003.
}
